Split SingleResponsibilityPrincipleManager class
