+++ src/main/java/edu/agh/wfiis/solid/srp/example2/CourseworkGenerator.java
@@ -1,0 +1,10 @@
+package edu.agh.wfiis.solid.srp.example2;
+
+public class CourseworkGenerator {
+
+	public String generateCourseworkForStudents() {
+        /* some magic happens here, irrelevant from this example perspective...*/
+        return "very difficult coursework";
+    }
+	
+}
