+++ src/main/java/edu/agh/wfiis/solid/srp/example2/SingleResponsibilityPrincipleDescriptionProvider.java
@@ -1,0 +1,10 @@
+package edu.agh.wfiis.solid.srp.example2;
+
+public class SingleResponsibilityPrincipleDescriptionProvider implements SolidRulesDescriptionProvider {
+
+	@Override
+	public String provideDescription() {
+        return "SRP description";
+    }
+
+}
