+++ src/main/java/edu/agh/wfiis/solid/srp/example2/SingleResponsibilityPrincipleRefactoringTool.java
@@ -1,0 +1,10 @@
+package edu.agh.wfiis.solid.srp.example2;
+
+class SingleResponsibilityPrincipleRefactoringTool {
+
+	public String refactorCode(String badCode) {
+        /* some magic happens here, irrelevant from this example perspective...*/
+        return "code refactored using SRP";
+    }
+	
+}
