+++ src/main/java/edu/agh/wfiis/solid/srp/example2/SolidRulesDescriptionProvider.java
@@ -1,0 +1,7 @@
+package edu.agh.wfiis.solid.srp.example2;
+
+public interface SolidRulesDescriptionProvider {
+
+	public String provideDescription();
+
+}
